Revert the Object folder rename, keep only the class
