+++ composer.json
@@ -1,5 +1,5 @@
 {
-    "name": "webonyx/omlex",
+    "name": "excelwebzone/omlex",
     "description": "Lightweight oEmbed for PHP 5.3",
     "keywords": ["oembed"],
     "homepage": "https://github.com/excelwebzone/Omlex",
