+++ lib/Omlex/OmlexObject.php
@@ -12,7 +12,7 @@
 namespace Omlex;
 
 use Omlex\Exception\NoSupportException;
-use Omlex\Exception\OmlexObjectException;
+use Omlex\Exception\ObjectException;
 
 /**
  * Base class for consuming Omlex objects.
@@ -36,17 +36,17 @@ abstract class OmlexObject
     /**
      * Create an Omlex object from result
      *
-     * @param object $object Raw object returned from API
+     * @param OmlexObject $object Raw object returned from API
      *
-     * @return object Instance of object driver
+     * @return OmlexObject Instance of object driver
      *
-     * @throws OmlexObjectException    On object errors
+     * @throws ObjectException    On object errors
      * @throws NoSupportException When object type is not supported or unknown
      */
     static public function factory($object)
     {
         if (!isset($object->type)) {
-            throw new OmlexObjectException('The object has no type.');
+            throw new ObjectException('The object has no type.');
         }
 
         $type = (string)$object->type;
@@ -54,9 +54,9 @@ abstract class OmlexObject
             throw new NoSupportException(sprintf('The object type "%s" is unknown or invalid.', $type));
         }
 
-        $class = '\\Omlex\\OmlexObject\\'.self::$types[$type];
+        $class = '\\Omlex\\Object\\'.self::$types[$type];
         if (!class_exists($class)) {
-            throw new OmlexObjectException(sprintf('The object class "%s" is invalid or not found.', $class));
+            throw new ObjectException(sprintf('The object class "%s" is invalid or not found.', $class));
         }
 
         $instance = new $class($object);
